made a lot of commands server specific

+++ events/guildMemberRemove.js
@@ -1,4 +1,5 @@
 module.exports = (bot, member) => {
+    if (member.guild.id !== bot.config.guildID) return;
     var screening = member.guild.channels.find('name', 'screening-lobby');
     screening.send(`${member.user.username} has left the server!`)
     .then(msg => setTimeout(function() {msg.delete()}, 60000));
